mtcp_restart_nolibc.c in dmtcp-unstable was modified to use mmapfile instead of readfile when restoring a checkpoint image _if_ the checkpoint image had not been compressed.  mmap should be faster than read, which is important for case of frequent restarts in reversible debugger.
This should only affect you if you're not using gzip.  If you suspect
it's causing a bug for you, change conditionalization
at mtcp_restart_no_libc:389 (unstable branch) from "#if 1" to #if 0".

The remaining changes are primarily to port some bug fixes of main branch
into unstable branch.  In particular, this should now fix the "bash bug"
for the unstable branch, also.


git-svn-id: svn://svn.code.sf.net/p/dmtcp/code/trunk@364 ba6dc0ae-232f-0410-883f-9ebac64fc565