MTCP Now uses DMTCP Allocator for malloc/free events.
MTCP now used dmtcp malloc/free funcptrs that are passed via
mtcp_init_dmtcp_info(). This is to avoid deadlocks that can happend when
a user thread is SUSPENDED inside libc::malloc and MTCP is trying to
malloc/free memory.

EDIT: This revision introduces a bug which has been fixed in rev 1248.


git-svn-id: svn://svn.code.sf.net/p/dmtcp/code/trunk@1243 ba6dc0ae-232f-0410-883f-9ebac64fc565