Re-organize some of PTRACE specific code [1/3]
This is a part of a series of changes to better reorganize the PTRACE
specific code. The higher goal here is to be able to create a ptrace
module which can be shipped/packaged. In order to do this, we want to
have the code organized in such a way that the calls from DMTCP code to
PTRACE specific code are as few as possible. This has to be true for
both dmtcp/src and mtcp/. Once we can reduce the number of these
hooks/function calls to a reasonable number, we can create MTCP specific
events and (may be) add a service similar to mtcp_process_event(). Of
course, we can always piggyback on dmtcp_process_event() but it might be
better to have a clear separation.

- Move Ptrace code from mtcpinterface.cpp to ptrace.cpp as needed.
- This change doesn't intend to make any logical/alogorithmic changes.

git-svn-id: svn://svn.code.sf.net/p/dmtcp/code/trunk@1293 ba6dc0ae-232f-0410-883f-9ebac64fc565