Allow child process to join computation after SUSPEND msg broadcast
If the child process is created after the coordinator has broadcasted
SUSPEND message, the newly created process should be allowed to join the
computation along with its parent.

The way we do this is by creating a new connection with the coordinator
before calling the _real_fork(). The parent process then closes this new
connection, whereas, the child process uses this connection as its
coordinator connection.

git-svn-id: svn://svn.code.sf.net/p/dmtcp/code/trunk@1362 ba6dc0ae-232f-0410-883f-9ebac64fc565