Move part of mtcp_set_callbacks() into mtcp_set_dmtcp_callbacks()
This is done to keep the mtcp api simpler. OpenMPI and other projects
that use MTCP directly without going through DMTCP, depend on a stable
and consistent MTCP API. This involves a consistent signature of
mtcp_set_callbacks(). Everytime we make any change to it, we modify
the API and need to update library version etc.

In order to avoid all this, we have now split the mtcp_set_callbacks()
function into two:
1. mtcp_set_callbacks(): This contains five core callbacks that anyone
   using MTCP might benefit from. These are sleep-between-ckpt, preCkpt,
   postCkpt, ckptFd, and write-ckpt-prefix.
2. mtcp_set_dmtcp_callbacks(): This is used exclusively be DMTCP to
   setup other callbacks.

git-svn-id: svn://svn.code.sf.net/p/dmtcp/code/trunk@1409 ba6dc0ae-232f-0410-883f-9ebac64fc565