Add holds_any_locks callback
(This work is based on the patch provided by Ana).

This changeset introduces a new callback holds_any_locks from MTCP to
DMTCP. This callback is useful only for the ptrace module currently, but
may be used for other stuff as well.

This is invoked as the first thing in stopthisthread() routine,
which is the signal handler for CKPT signal, to check if the current
thread is holding any of the wrapperExecLock or threadCreationLock. If
the thread is holding any of these locks, we return from the signal
handler and wait for the thread to release the lock. Once the thread has
release the last lock, it will send itself the CKPT signal and will
return to the signal handler and will proceed normally.

git-svn-id: svn://svn.code.sf.net/p/dmtcp/code/trunk@1410 ba6dc0ae-232f-0410-883f-9ebac64fc565