Second commit related to Torque PBS support in DMTCP Correct processing of service Torque PBS files is added. There is 2 kinds of importand Torque files: 1. Standart output and error files (sometimes the same) 2. PBS node file
For the first type we need to save content of output and copy it into 
newly created output file. This provides feel of continued computations.

For the second file we also need save content of initial node file into
checkpoint and copy it to temporal file that will represent initial content to
application. 



git-svn-id: svn://svn.code.sf.net/p/dmtcp/code/trunk@1469 ba6dc0ae-232f-0410-883f-9ebac64fc565