Introduce libdlLock for dlopen/dlclose wrappers.
While dlopen is loading a shared library, if a constructor for a static
object calls fork(), the process could deadlock because the dlopen()
wrapper had acquired the wrapperExecLock while the fork() wrapper
wants toe acquire the same lock with writer permissions. The fix is to
create an extra lock specifically for dlopen/dlclose. See comments in
miscwrappers.cpp for more details.

git-svn-id: svn://svn.code.sf.net/p/dmtcp/code/trunk@1901 ba6dc0ae-232f-0410-883f-9ebac64fc565