PTY: Bug fix for CTTY in multiple process trees.
It is possible to have two different connection-ids for a pre-existing
CTTY in two or more different process trees. In this case, only one of
the several process trees would be able to acquire a lock on the
underlying fd.  The send-receive fd logic fails in this case due to
different connection-ids.  Therefore, we let every process do a
postRestart to reopen the CTTY.

TODO: A better fix would be to have a unique connection-id for each
pre-existing CTTY that is then used by all process trees.  It can be
implemented by using the SharedData area.

git-svn-id: svn://svn.code.sf.net/p/dmtcp/code/trunk@1947 ba6dc0ae-232f-0410-883f-9ebac64fc565