MTCP: Bug fix when running with huge environment.
The static string buffer used to store arguments and environ on restart.
On some systems, it fails to print the debug message during overflow and
causes the restarted process to hang. A proper fix would be to not rely
on the static size and come up with an alternate solution.

git-svn-id: svn://svn.code.sf.net/p/dmtcp/code/trunk@2210 ba6dc0ae-232f-0410-883f-9ebac64fc565