Jalib: Increase buffer size for getProgramArgs
getdelim() will call realloc if the supplied buffer is not big enough to
hold the returning string.  However, since we did a jalib::malloc,
libc:realloc() fails with invalid pointer. The temporary fix is to
increase the original buffer. Eventually, we would need to replace
getdelim with our own version.

git-svn-id: svn://svn.code.sf.net/p/dmtcp/code/trunk@2796 ba6dc0ae-232f-0410-883f-9ebac64fc565