xlib_wrappers: Fix saving of null display name in XOpenDisplay
The code to save the display name in XOpenDisplay() was incorrect.
When an application passes a NULL pointer the string would get
truncated. The patch fixes this problem by explicitly typecasting
the char pointer to void* to just save the address instead of the
string. A cleaner solution would involve malloc-and-copy of a
non-null string argument.